karmaquest-backend/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-CORS==5.0.0
Flask-JWT-Extended==4.7.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
bcrypt==4.2.1
werkzeug==3.1.3
sqlalchemy==2.0.36
tensorflow==2.20.0
opencv-python-headless==4.12.0.88
Pillow==12.0.0
numpy==2.2.6
pytest==8.3.4
pytest-flask==1.3.0
pytest-cov==6.0.0
faker==33.1.0
